增加新闻发布，及前端模块

+++ modules/news/news-model/src/main/java/com/lebaoxun/modules/news/entity/NewsEntity.java
@@ -20,8 +20,8 @@ public class NewsEntity implements Serializable {
 	/**
 	 * 主键
 	 */
-@TableId
-	private Integer id;
+	@TableId
+	private long id;
 	/**
 	 * 作者ID
 	 */
@@ -78,13 +78,13 @@ public class NewsEntity implements Serializable {
 	/**
 	 * 设置：主键
 	 */
-	public void setId(Integer id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	/**
 	 * 获取：主键
 	 */
-	public Integer getId() {
+	public long getId() {
 		return id;
 	}
 	/**
